Cal 67 - Modify Chipping Service to Chip by Pixel Instead of Lon/Lat. (#68)
* replaced previous calculations with linear algebra.

* CAL-67 Chipping Service Assumes Base Image Is North Aligned

This change is a re-implementation of the original ChipServiceImpl that no longer assumes the image is north-aligned.  The interface
is unchanged.  The new implementation uses a matrix library and linear transormation to calculate the pixels in the sub-image.

* CAL-67 Add chipping by pixel to the chipping service

This changes adds a method to the ChipService to chip by pixel.  There are also fixes to the original chip method
implementation and changes to the Chip MetacardTransformer so that it now uses the chip-by-pixel method.