CAL-284 Update imaging-app bundles to handle the separation of imaging-nitf into bundles. (#292)
This change removes embedded dependencies that were necessary prior to the separation of imaging-nitf-fluent and imaging-nitf-core into api/impl
modules. It also includes code updates necessary to allow the imaging-app bundles to work with api-changes in 0.8 of imaging-nitf.

The NitfImageTransformer filters out coordinates which aren't in decimal degrees or degrees/minutes/seconds format. This change allows
UTM and MGRS values to be converted as well.

Updated imaging-nitf version to 0.8.1.